Enhance orders API, product model, and user admin
Added custom actions to OrderViewSet for employee and client-specific order access and state changes. Updated product model and migration to make description optional. Improved user admin with a custom UserAdmin and included 'role' in the user serializer. Minor fix in entrypoint.sh for makemigrations and set basename in orders router.